DTO 추가해서 Service 구현

+++ src/main/java/com/study/delivery/food/FoodService.java
@@ -3,6 +3,7 @@ package com.study.delivery.food;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -12,14 +13,32 @@ public class FoodService {
     MapRepository mapRepository;
     
     // 음식 목록 조회
-    public List<Food> getFoods() {
+    public List<FoodDto> getFoods() {
 
-        return mapRepository.getAll();
+        //전체 데이터 가져오기
+        List<Food> foods = mapRepository.getAll();
+        List<FoodDto> dtos = new ArrayList<FoodDto>();
+
+        //Entity->DTO
+        for(int i = 0; i< foods.size(); i++){
+            Food target = foods.get(i);
+            FoodDto dto = FoodDto.createFoodDto(target);
+            dtos.add(dto);
+        }
+
+        return dtos;
     }
 
     // 음식 상세 조회
-    public Food getFoodById(Integer foodId) {
-        return mapRepository.getById(foodId);
+    public FoodDto getFoodById(Integer foodId) {
+
+        //데이터 상세 조회
+        Food food = mapRepository.getById(foodId);
+
+        //Entity -> DTO
+        FoodDto dto = FoodDto.createFoodDto(food);
+
+        return dto;
     }
 
 
@@ -37,13 +56,28 @@ public class FoodService {
 
 
     // 음식 수정
-    public void updateFood(Food food) {
-        mapRepository.update(food);
+    public void updateFood(Integer foodId, FoodDto dto) {
+
+        //수정할 데이터 Entity 변환
+        Food data = dto.toEntity();
+
+        //해당아이디 게시글 확인 후 수정
+        Food target = mapRepository.getById(foodId);
+        if(target != null){
+            mapRepository.update(data);
+        }
+        
     }
 
     // 음식 삭제
     public void deleteFood(Integer foodId) {
-        mapRepository.delete(foodId);
+        
+        //해당아이디 게시글 확인 후 삭제
+        Food target = mapRepository.getById(foodId);
+        if (target != null) {
+            mapRepository.delete(foodId);
+        }
+      
     }
